Rename `frequency_compounding` module to `frequency`.
